+++ src/Usages/Data/ResolvedEntityUsage.php
@@ -59,16 +59,4 @@ final class ResolvedEntityUsage
     {
         return $this->url;
     }
-
-
-    /**
-     */
-    public function toArray () : array
-    {
-        return [
-            "name" => $this->labels,
-            "type" => $this->type,
-            "url" => $this->url,
-        ];
-    }
 }